# Conflicts:
#	src/main/java/vn/edu/fpt/petworldplatform/service/PetService.java
#	src/main/resources/application.properties

+++ src/main/java/vn/edu/fpt/petworldplatform/repository/AppointmentRepository.java
@@ -14,7 +14,7 @@ public interface AppointmentRepository extends JpaRepository<Appointment, Intege
 
     List<Appointment> findByCustomerIdOrderByAppointmentDateDesc(Integer customerId);
 
-    List<Appointment> findByCustomerIdAndStatusInOrderByAppointmentDateDesc(Long customerId, List<String> statuses);
+    List<Appointment> findByCustomerIdAndStatusInOrderByAppointmentDateDesc(Integer customerId, List<String> statuses);
 
     long countByAppointmentDateAndStatusNot(LocalDateTime appointmentDate, String excludedStatus);
